+++ src/main/scala/wdlTools/eval/Stdlib.scala
@@ -293,6 +293,7 @@ case class Stdlib(paths: EvalPaths,
       Builtins.Range -> range,
       Builtins.Transpose -> transpose,
       Builtins.Zip -> zip,
+      Builtins.Unzip -> unzip,
       Builtins.Cross -> cross,
       Builtins.Flatten -> flatten,
       Builtins.Prefix -> prefix,
@@ -345,6 +346,7 @@ case class Stdlib(paths: EvalPaths,
       Builtins.Range -> range,
       Builtins.Transpose -> transpose,
       Builtins.Zip -> zip,
+      Builtins.Unzip -> unzip,
       Builtins.Cross -> cross,
       Builtins.Flatten -> flatten,
       Builtins.Prefix -> prefix,
@@ -823,6 +825,7 @@ case class Stdlib(paths: EvalPaths,
   // This function returns the basename of a file path passed to it: basename("/path/to/file.txt") returns "file.txt".
   // Also supports an optional parameter, suffix to remove: basename("/path/to/file.txt", ".txt") returns "file".
   // since: draft-2
+  // supports regular expression for second parameter since 1.1
   private def basename(ctx: FunctionContext): V_String = {
     ctx.assertNumArgs(required = 1, optional = true)
     val filePath = ctx.args.head match {
@@ -833,7 +836,15 @@ case class Stdlib(paths: EvalPaths,
     }
     val filename = Paths.get(filePath).getFileName.toString
     if (ctx.args.size == 2) {
-      V_String(filename.stripSuffix(getWdlString(ctx.args(1), ctx.loc)))
+      // make sure the pattern only matches at the end of the filename
+      val suffixPattern = getWdlString(ctx.args(1), ctx.loc) match {
+        case s if s.endsWith("$") => s
+        case s                    => s"${s}$$"
+      }
+      val re2Pattern = Pattern.compile(suffixPattern, 0)
+      val re2Matcher = re2Pattern.matcher(filename)
+      val result = re2Matcher.replaceFirst("")
+      V_String(result)
     } else {
       V_String(filename)
     }
@@ -1237,12 +1248,7 @@ case class Stdlib(paths: EvalPaths,
     } else {
       nBytes
     }
-    if (result.isExactDouble) {
-      V_Float(result.toDouble)
-    } else {
-      throw new EvalException(s"size: overflow error - cannot represent size ${result} as a float",
-                              ctx.loc)
-    }
+    V_Float(result.toDouble)
   }
 
   // Integer length(Array[X])
@@ -1285,6 +1291,18 @@ case class Stdlib(paths: EvalPaths,
     V_Array(xArray.zip(yArray).map {
       case (x, y) => V_Pair(x, y)
     })
+  }
+
+  // Pair[Array[X], Array[Y]] unzip(Array[Pair[X, Y]])
+  //
+  // since: 1.1
+  private def unzip(ctx: FunctionContext): V_Pair = {
+    val array: Vector[V] = getWdlVector(ctx.getOneArg, ctx.loc)
+    val (x, y) = array.map {
+      case V_Pair(l, r) => (l, r)
+      case other        => throw new EvalException(s"unzip: invalid array element ${other}")
+    }.unzip
+    V_Pair(V_Array(x), V_Array(y))
   }
 
   // Array[Pair(X,Y)] cross(Array[X], Array[Y])